separate redis instance to store rooms
