+++ soupware-orch/.env
@@ -1,5 +1,6 @@
 NATS=localhost
 MEDIA_NODES_STORE_ADDR=localhost:6379
+ROOMS_STORE_ADDR=localhost:6380
 SOUPWARE_JWT_SECRET=dev
 
 SOUPWARE_WEBHOOK=http://localhost:3000/_soupware/webhook
