+++ soupware-orch/src/node-manager/node-manager.service.ts
@@ -1,17 +1,10 @@
-import {
-  Injectable,
-  OnApplicationShutdown,
-  OnModuleInit,
-} from '@nestjs/common';
+import { Injectable, OnModuleInit } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { MediaNodeLoad } from '@soupware/internals';
 import Redis from 'ioredis';
 
-const PREFIX_ROOM_PIPES = 'room_pipes_';
-const PREFIX_NODES_FOR_ROOM = 'nodes_for_';
-
 @Injectable()
-export class NodeManagerService implements OnModuleInit, OnApplicationShutdown {
+export class NodeManagerService implements OnModuleInit {
   private redis: Redis;
 
   constructor(private configService: ConfigService) {}
@@ -20,39 +13,8 @@ export class NodeManagerService implements OnModuleInit, OnApplicationShutdown {
     this.redis = new Redis(this.configService.get('MEDIA_NODES_STORE_ADDR'));
   }
 
-  async onApplicationShutdown() {
-    if (process.env.NODE_ENV !== 'production') {
-      await this.redis.flushdb();
-    }
-  }
-
   async updateNodeLoad(data: MediaNodeLoad) {
-    console.log('received node load', data);
     await this.redis.hmset(data.id, data);
-  }
-
-  async addNodeForRoom(room: string, node: string) {
-    await this.redis.sadd(PREFIX_NODES_FOR_ROOM + room, node);
-  }
-
-  async getSendNodesFor(room: string) {
-    const nodes = await this.redis.smembers(PREFIX_NODES_FOR_ROOM + room);
-
-    return nodes.filter((node) => node.startsWith('SEND'));
-  }
-
-  async addRoomPipe(room: string, pipedRecvNodeId: string) {
-    await this.redis.sadd(PREFIX_ROOM_PIPES + room, pipedRecvNodeId);
-  }
-
-  async getPipedNodeIds(room: string) {
-    return this.redis.smembers(PREFIX_ROOM_PIPES + room);
-  }
-
-  async isRoomPipedTo(room: string, recvNodeId: string) {
-    return (
-      (await this.redis.sismember(PREFIX_ROOM_PIPES + room, recvNodeId)) === 1
-    );
   }
 
   async delNode(id: string, kind: string) {
