+++ soupware-orch/src/room/room.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { RoomService } from './room.service';
+
+@Module({
+  imports: [],
+  providers: [RoomService],
+  controllers: [],
+  exports: [RoomService],
+})
+export class RoomModule {}
